Refactor  : 확장될 가능성 없는 서비스 계층 인터페이스 -> 구현체 리팩토링
