+++ src/main/java/com/flab/delivery/service/AddressService.java
@@ -2,13 +2,52 @@ package com.flab.delivery.service;
 
 import com.flab.delivery.dto.address.AddressDto;
 import com.flab.delivery.dto.address.AddressRequestDto;
+import com.flab.delivery.exception.AddressException;
+import com.flab.delivery.mapper.AddressMapper;
+import com.flab.delivery.mapper.UserAddressMapper;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
 
 import java.util.List;
 
-public interface AddressService {
-    void addAddress(AddressRequestDto addressRequestDto, String userId);
+@Service
+@RequiredArgsConstructor
+public class AddressService {
 
-    List<AddressDto> getAllAddress(String userId);
+    private final AddressMapper addressMapper;
+    private final UserAddressMapper userAddressMapper;
 
-    void removeAddress(Long userAddressId, String userId);
+    public void addAddress(AddressRequestDto addressRequestDto, String userId) {
+
+        Long findAddressId = addressMapper.findIdByTownName(addressRequestDto.getTownName());
+
+        if (findAddressId == null) {
+            throw new AddressException("존재하지 않는 주소 입니다.");
+        }
+
+        if (addressRequestDto.getAlias() == null) {
+            addressRequestDto.changeAlias(addressRequestDto.getDetailAddress());
+        }
+
+        userAddressMapper.addAddress(addressRequestDto, findAddressId, userId);
+
+    }
+
+    public List<AddressDto> getAllAddress(String userId) {
+
+        return userAddressMapper.findAllByUserId(userId);
+    }
+
+    public void removeAddress(Long userAddressId, String userId) {
+
+        if (!userAddressMapper.existsById(userAddressId)) {
+            throw new AddressException("존재하지 않는 주소 입니다.");
+        }
+
+        userAddressMapper.deleteById(userAddressId);
+    }
+
+    public void selectAddress(Long id, String userId) {
+
+    }
 }
